7 - text to speech/requirements.txt
gtts-2.1.1 
gtts-token-1.1.3